combine Router and Renderer into Client
