+++ packages/kit/src/runtime/client/singletons.js
@@ -1,19 +1,11 @@
-/**
- * The router is nullable, but not typed that way for ease-of-use
- * @type {import('./router').Router}
- */
-export let router;
-
-/** @type {import('./renderer').Renderer} */
-export let renderer;
+/** @type {import('./client').Client} */
+export let client;
 
 /**
  * @param {{
- *   router: import('./router').Router?;
- *   renderer: import('./renderer').Renderer;
+ *   client: import('./client').Client;
  * }} opts
  */
 export function init(opts) {
-	router = /** @type {import('../client/router').Router} */ (opts.router);
-	renderer = opts.renderer;
+	client = opts.client;
 }
