+++ packages/kit/src/runtime/client/start.js
@@ -1,8 +1,7 @@
 import Root from '__GENERATED__/root.svelte';
 // @ts-expect-error - doesn't exist yet. generated by Rollup
 import { routes, fallback } from '__GENERATED__/manifest.js';
-import { Router } from './router.js';
-import { Renderer } from './renderer.js';
+import { Client } from './client.js';
 import { init } from './singletons.js';
 import { set_paths } from '../paths.js';
 
@@ -26,29 +25,23 @@ import { set_paths } from '../paths.js';
  * }} opts
  */
 export async function start({ paths, target, session, route, spa, trailing_slash, hydrate }) {
-	const renderer = new Renderer({
+	const client = new Client({
 		Root,
 		fallback,
 		target,
-		session
+		session,
+		base: paths.base,
+		routes,
+		trailing_slash
 	});
 
-	const router = route
-		? new Router({
-				base: paths.base,
-				routes,
-				trailing_slash,
-				renderer
-		  })
-		: null;
-
-	init({ router, renderer });
+	init({ client });
 	set_paths(paths);
 
-	if (hydrate) await renderer.start(hydrate);
-	if (router) {
-		if (spa) router.goto(location.href, { replaceState: true }, []);
-		router.init_listeners();
+	if (hydrate) await client.start(hydrate);
+	if (route) {
+		if (spa) client.goto(location.href, { replaceState: true }, []);
+		client.init_listeners();
 	}
 
 	dispatchEvent(new CustomEvent('sveltekit:start'));
